+++ src/app/admin/components/certifiedProduct/listing/edit.controller.js
@@ -12,13 +12,10 @@
         vm.addNewValue = utilService.addNewValue;
         vm.attachModel = attachModel;
         vm.cancel = cancel;
-        vm.directCertsDirective = directCertsDirective;
         vm.disabledParent = disabledParent;
         vm.disabledStatus = disabledStatus;
         vm.extendSelect = utilService.extendSelect;
-        vm.prep = prep;
         vm.requiredIcsCode = requiredIcsCode;
-        vm.registerCerts = registerCerts;
         vm.save = save;
         vm.willCauseSuspension = willCauseSuspension;
 
@@ -77,12 +74,6 @@
             $uibModalInstance.dismiss('cancelled');
         }
 
-        function directCertsDirective () {
-            angular.forEach(vm.handlers, function (handler) {
-                handler();
-            });
-        }
-
         function disabledParent (listing) {
             var ret = false;
             ret = ret || vm.cp.chplProductNumber === listing.chplProductNumber;
@@ -96,27 +87,12 @@
             return ((name === 'Pending' && vm.workType === 'manage') || (name !== 'Pending' && vm.workType === 'confirm'));
         }
 
-        function prep () {
-            vm.directCertsDirective();
-            $timeout(vm.save, 1000);
-        }
-
         function requiredIcsCode () {
             var code = vm.cp.ics.parents
                 .map(function (item) { return parseInt(item.chplProductNumber.split('.')[6], 10); })
                 .reduce(function (max, cur) { return Math.max(max, cur); }, -1)
                 + 1;
             return (code > 9 || code < 0) ? '' + code : '0' + code;
-        }
-
-        function registerCerts (handler) {
-            vm.handlers.push(handler);
-            var removeHandler = function () {
-                vm.handlers = vm.handlers.filter(function (aHandler) {
-                    return aHandler !== handler;
-                });
-            };
-            return removeHandler;
         }
 
         function save () {